divided plotting and histogram filling utils in separate modules
