+++ plotting/plot_utils.py
@@ -161,22 +161,6 @@ def combineYears(inplots, tag='QCD_HT', years=['2018','2017','2016']):
         initialize = False
     
     return outPlots
-
-# load hdf5 with pandas
-def h5load(ifile, label):
-    try:
-        with pd.HDFStore(ifile, 'r') as store:
-            try:
-                data = store[label] 
-                metadata = store.get_storer(label).attrs.metadata
-                return data, metadata
-
-            except KeyError:
-                print("No key",label,ifile)
-                return 0, 0
-    except:
-        print("Some error occurred", ifile)
-        return 0, 0
  
 def check_proxy(time_min=100):
     """
@@ -213,28 +197,6 @@ def check_proxy(time_min=100):
         
     return lifetime
 
-def getXSection(dataset, year, SUEP=False):
-    xsection = 1
-    if not SUEP:
-        with open('../data/xsections_{}.json'.format(year)) as file:
-            MC_xsecs = json.load(file)
-            try:
-                xsection *= MC_xsecs[dataset]["xsec"]
-                xsection *= MC_xsecs[dataset]["kr"]
-                xsection *= MC_xsecs[dataset]["br"]
-            except:
-                print("WARNING: I did not find the xsection for that MC sample. Check the dataset name and the relevant yaml file")
-                return 1
-    else:
-        with open('../data/xsections_{}_SUEP.json'.format(year)) as file:
-            MC_xsecs = json.load(file)
-            try:
-                xsection *= MC_xsecs[dataset]
-            except:
-                print("WARNING: I did not find the xsection for that MC sample. Check the dataset name and the relevant yaml file")
-                return 1
-    return xsection
-
 def apply_binwise_scaling(h_in, bins, scales, dim='x'):
     """
     Apply scales to bins of a particular histogram.
@@ -257,38 +219,6 @@ def apply_binwise_scaling(h_in, bins, scales, dim='x'):
         h[:,:] = np.stack([new_z, new_z], axis=-1)
     return h
 
-def make_selection(df, variable, operator, value, apply=True):
-    """
-    Apply a selection on DataFrame df based on on the df column'variable'
-    using the 'operator' and 'value' passed as arguments to the function.
-    Returns the resulting DataFrame after the operation is applied.
-    
-    df: input dataframe.
-    variable: df column.
-    operator: see code below.
-    value: value to cut variable on using operator.
-    apply: toggles whether the selection is applied to the dataframe, or
-    whether a list of booleans is returned matching the indices that
-    passed and failed the selection.
-    """
-    if operator in ["greater than","gt",">"]:
-        if apply: return df.loc[(df[variable] > value)]
-        else: return (df[variable] > value)
-    if operator in ["greater than or equal to", ">="]:
-        if apply: return df.loc[(df[variable] >= value)]
-        else: return (df[variable] >= value)
-    elif operator in ["less than", "lt", "<"]:
-        if apply: return df.loc[(df[variable] < value)]
-        else: return (df[variable] < value)
-    elif operator in ["less than or equal to", "<="]:
-        if apply: return df.loc[(df[variable] <= value)]
-        else: return (df[variable] <= value)
-    elif operator in ["equal to", "eq", "=="]:
-        if apply: return df.loc[(df[variable] == value)]
-        else: return (df[variable] == value)
-    else:
-        sys.exit("Couldn't find operator requested " + operator)
-
 # function to load files from pickle
 def openpkl(infile_name):
     plots={}
@@ -299,170 +229,6 @@ def openpkl(infile_name):
             except EOFError:
                 break
     return plots
-
-def apply_scaling_weights(df, scaling_weights,
-    x_var_regions, y_var_regions,
-    regions = "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
-    x_var = 'SUEP_S1_CL',
-    y_var = 'SUEP_nconst_CL',
-    z_var = 'ht'):
-    """
-    df: input DataFrame to scale
-    *_var_regions: x/y ABCD regions
-    scaling_weights: nested dictionary, region x (bins or ratios)
-    regions: string of ordered regions, used to apply corrections
-    *_var: x/y are of the ABCD plane, z of the scaling histogram
-    """
-    
-    x_var_regions = abcd['x_var_regions']
-    y_var_regions = abcd['y_var_regions']
-    iRegion = 0
-
-    # S1 regions
-    for i in range(len(x_var_regions)-1):
-        x_val_lo = x_var_regions[i]
-        x_val_hi = x_var_regions[i+1]
-
-        # nconst regions
-        for j in range(len(y_var_regions)-1):
-            y_val_lo = y_var_regions[j]
-            y_val_hi = y_var_regions[j+1]
-
-            r = regions[iRegion]
-
-            # from the weights
-            bins = weights[r]['bins']
-            ratios = weights[r]['ratios']
-
-            # ht bins
-            for k in range(len(bins)-1):
-                z_val_lo = bins[k]
-                z_val_hi = bins[k+1]
-                ratio = ratios[k]
-
-                zslice = (df[z_var] >= z_val_lo) & (df[z_var] < z_val_hi)
-                yslice = (df[y_var] >= y_val_lo) & (df[y_var] < y_val_hi)
-                xslice = (df[x_var] >= x_val_lo) & (df[x_var] < x_val_hi)
-
-                df.loc[xslice & yslice & zslice, 'event_weight'] *= ratio
-
-            iRegion += 1
-    return df
-
-def auto_fill(df, output, abcd, label_out, isMC=False, do_abcd=False):
-    
-    input_method = abcd['input_method']
-
-    #####################################################################################
-    # ---- Fill Histograms
-    # Automatically fills all histograms that are declared in the output dict.
-    #####################################################################################
-    
-    # 1. fill the distributions as they are saved in the dataframes
-    # 1a. Plot event wide variables
-    plot_labels = [key for key in df.keys() if key+"_"+label_out in list(output.keys())]  
-    for plot in plot_labels: output[plot+"_"+label_out].fill(df[plot], weight=df['event_weight']) 
-    
-    # 1b. Plot method variables
-    plot_labels = [key for key in df.keys() if key.replace(input_method, label_out) in list(output.keys()) and key.endswith(input_method)]
-    for plot in plot_labels: 
-        if not plot.endswith(input_method): continue
-        output[plot.replace(input_method, label_out)].fill(df[plot], weight=df['event_weight'])  
-    # FIXME: plot ABCD 2d
-    
-    # 2. fill some 2D distributions  
-    keys = list(output.keys())
-    keys_2Dhists = [k for k in keys if '2D' in k]
-    for key in keys_2Dhists:
-        if not key.endswith(label_out): continue
-        string = key[len("2D")+1:-(len(label_out)+1)] # cut out "2D_" and output label
-        var1 = string.split("_vs_")[0]
-        var2 = string.split("_vs_")[1]
-        if var1 not in list(df.keys()): var1 += "_" + input_method
-        if var2 not in list(df.keys()): var2 += "_" + input_method
-        output[key].fill(df[var1], df[var2], weight=df['event_weight'])
-
-    # 3. divide the dfs by region
-    if do_abcd:
-        regions = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-        xvar = abcd['xvar']
-        yvar = abcd['yvar']
-        xvar_regions = abcd['xvar_regions']
-        yvar_regions = abcd['yvar_regions']
-        iRegion = 0
-        for i in range(len(xvar_regions)-1):
-            x_val_lo = xvar_regions[i]
-            x_val_hi = xvar_regions[i+1]
-
-            for j in range(len(yvar_regions)-1):
-                y_val_lo = yvar_regions[j]
-                y_val_hi = yvar_regions[j+1]
-                
-                x_cut = (make_selection(df, xvar, '>=', x_val_lo, False) & make_selection(df, xvar, '<', x_val_hi, False))
-                y_cut = (make_selection(df, yvar, '>=', y_val_lo, False) & make_selection(df, yvar, '<', y_val_hi, False))
-                df_r = df.loc[(x_cut & y_cut)]
-
-                # double check the region is defined correctly
-                assert df_r[(df_r[xvar] > float(x_val_hi)) | (df_r[xvar] < float(x_val_lo))].shape[0] == 0
-
-                r = regions[iRegion] + "_"
-                iRegion += 1
-
-                # double check blinding
-                if iRegion == (len(xvar_regions)-1)*(len(yvar_regions)-1) and not isMC:
-                    if df_r.shape[0] > 0: 
-                        sys.exit(label_out+": You are not blinding correctly! Exiting.")
-
-                # 3a. Plot event wide variables
-                plot_labels = [key for key in df_r.keys() if r+key+"_"+label_out in list(output.keys())]   # event wide variables
-                for plot in plot_labels: 
-                    output[r+plot+"_"+label_out].fill(df_r[plot], weight=df_r['event_weight']) 
-                # 3b. Plot method variables
-                plot_labels = [key for key in df_r.keys() if r+key.replace(input_method, label_out) in list(output.keys())]  # method vars
-                for plot in plot_labels: 
-                    if not plot.endswith(input_method): continue
-                    
-                    output[r+plot.replace(input_method, label_out)].fill(df_r[plot], weight=df_r['event_weight'])  
-
-def apply_normalization(plots, xsection, total_gensumweight):
-    if total_gensumweight > 0.0:
-        for plot in list(output.keys()): plots[plot] = plots[plot]*xsection/total_gensumweight
-    else:
-        logging.warning("Total gensumweight is 0")
-    return plots
-
-def get_track_killing_config(config):
-    new_config = {}
-    for label_out, config_out in config.items():
-        label_out_new = label_out+"_track_down"
-        new_config[label_out_new] = deepcopy(config[label_out])
-        new_config[label_out_new]['input_method'] += "_track_down"
-        new_config[label_out_new]['xvar'] += "_track_down"
-        new_config[label_out_new]['yvar'] += "_track_down"
-        for iSel in range(len(new_config[label_out_new]['SR'])):
-            new_config[label_out_new]['SR'][iSel][0] += "_track_down"
-        for iSel in range(len(new_config[label_out_new]['selections'])):
-            if new_config[label_out_new]['selections'][iSel][0] in ['ht', 'ngood_ak4jets']: continue
-            new_config[label_out_new]['selections'][iSel][0] += "_track_down"
-        
-    return new_config
-
-def get_jet_corrections_config(config, jet_corrections):
-    new_config = {}
-    for sys in jet_corrections: 
-        for label_out, config_out in config.items():
-            label_out_new = label_out+"_"+sys
-            new_config[label_out_new] = deepcopy(config[label_out])
-            for iSel in range(len(new_config[label_out_new]['selections'])):
-                if 'ht' == new_config[label_out_new]['selections'][iSel][0]:
-                    new_config[label_out_new]['selections'][iSel][0] += "_" + sys 
-    return new_config
-
-def read_in_weights(fweights):
-    w = np.load(fweights, allow_pickle=True)
-    scaling_weights = defaultdict(lambda: np.zeros(2))
-    scaling_weights.update(w.item())
-    return scaling_weights
 
 def plot1d(h, ax, label, color='default', lw=1):
